requirements.txt
djangorestframework==3.6.4
requests==2.18.4
